Index: configure.in =================================================================== - changed version to 1.19.1 - added DYNLINK variable. This is set to "" on i686 platforms and   "-fPIC" otherwise
Index: include/RNumber.h
Index: src/RNumber.C
===================================================================
- added new ctors for RNUMBER and marked all but one as explicit.
- added call to constructor where it was previously implicit in multiply().

Index: src/Random.C
===================================================================
- added call to constructor where it was previously implicit.

Index: tests/NumberTest.C
===================================================================
- added call to RNumber(long long ....) ctor.

Index: tests/ShiftTest.C
===================================================================
static_cast of constant number to help the compiler find the correct
ctor.

+++ tests/ShiftTest.C
@@ -7,6 +7,6 @@ using namespace std;
 using namespace rnumber;
 
 int main ( int argc, char ** argv) {
-  cout << "10 << rn 35" << ( 10 << RNumber(0x35)) << endl;
+  cout << "10 << rn 35" << ( 10 << RNumber(static_cast<unsigned int>(0x35))) << endl;
   cout << "10 >> rn 35" << ( 10 >> RNumber(0x35)) << endl;
 }
